+++ MASFoundation/Classes/_private_/services/access/MASAccessService.h
@@ -24,6 +24,7 @@ typedef NS_ENUM(NSInteger, MASAccessValueType)
 {
     MASAccessValueTypeUknonw = -1,
     MASAccessValueTypeAccessToken,
+    MASAccessValueTypeAuthenticatedUsername,
     MASAccessValueTypeRefreshToken,
     MASAccessValueTypeScope,
     MASAccessValueTypeTokenType,