+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ### Bug fixes
  
 - Adding listAttibutes to MASObject in MASFoundation framework after removing from MASStorage MASObject category.
+- Fix the bug where MASUser object's isAuthenticated property returns wrong value for other user objects retrieved from MASIdentityManagement.
 
 ### New features